Factorize batched tests.
