+++ tests/integration/gendil/test-batched-l2-error.cpp
@@ -4,9 +4,9 @@
 
 #include <gendil/gendil.hpp>
 
-#include <algorithm>
+#include "batched-cell-test-helpers.hpp"
+
 #include <array>
-#include <cmath>
 #include <iostream>
 
 #if !defined(GENDIL_USE_DEVICE)
@@ -22,86 +22,10 @@ int main()
 #else
 
 using namespace gendil;
+using namespace gendil::test;
 
 namespace
 {
-constexpr Real sentinel = -12345.25;
-
-template < typename T >
-struct DeviceBuffer
-{
-   HostDevicePointer< T > data;
-   GlobalIndex size;
-
-   DeviceBuffer( const GlobalIndex n, const T value )
-      : size( n )
-   {
-      AllocateHostPointer( size, data );
-      AllocateDevicePointer( size, data );
-      Fill( value );
-   }
-
-   DeviceBuffer( const DeviceBuffer & ) = delete;
-   DeviceBuffer & operator=( const DeviceBuffer & ) = delete;
-
-   ~DeviceBuffer()
-   {
-      FreeDevicePointer( data );
-      FreeHostPointer( data );
-   }
-
-   void Fill( const T value )
-   {
-      for ( GlobalIndex i = 0; i < size; ++i )
-      {
-         data.host_pointer[ i ] = value;
-      }
-      ToDevice( size, data );
-   }
-
-   void CopyToHost() const
-   {
-      ToHost( size, data );
-   }
-};
-
-Integer GetDeviceMaxThreadsPerBlock()
-{
-#if defined( GENDIL_USE_CUDA )
-   int device = 0;
-   cudaDeviceProp properties;
-   GENDIL_GPU_CHECK( cudaGetDevice( &device ) );
-   GENDIL_GPU_CHECK( cudaGetDeviceProperties( &properties, device ) );
-   return static_cast< Integer >( properties.maxThreadsPerBlock );
-#elif defined( GENDIL_USE_HIP )
-   int device = 0;
-   hipDeviceProp_t properties;
-   GENDIL_GPU_CHECK( hipGetDevice( &device ) );
-   GENDIL_GPU_CHECK( hipGetDeviceProperties( &properties, device ) );
-   return static_cast< Integer >( properties.maxThreadsPerBlock );
-#else
-   return 1;
-#endif
-}
-
-template < typename Layout, Integer BatchSize >
-bool LaunchConfigurationFits( const char * label )
-{
-   const Integer requested_threads =
-      Layout::GetNumberOfThreads() * BatchSize;
-   const Integer max_threads = GetDeviceMaxThreadsPerBlock();
-
-   if ( requested_threads > max_threads )
-   {
-      std::cout << "Skipping " << label
-                << ": requested " << requested_threads
-                << " threads per block, device limit is "
-                << max_threads << ".\n";
-      return false;
-   }
-   return true;
-}
-
 Vector MakeInputVector( const Integer size )
 {
    return Vector(
@@ -134,70 +58,6 @@ Real ComputeL2Error(
    return error;
 }
 
-bool CheckClose(
-   const char * label,
-   const Real observed,
-   const Real expected,
-   const Real tolerance )
-{
-   const Real abs_error = std::abs( observed - expected );
-   const Real scale = std::max( Real{ 1.0 }, std::abs( expected ) );
-   const Real scaled_error = abs_error / scale;
-
-   std::cout << label << " observed = " << observed
-             << ", expected = " << expected
-             << ", scaled error = " << scaled_error << '\n';
-
-   if (
-      !std::isfinite( observed ) ||
-      !std::isfinite( expected ) ||
-      !std::isfinite( scaled_error ) ||
-      scaled_error > tolerance )
-   {
-      std::cout << "FAILED: " << label << " exceeded tolerance "
-                << tolerance << ".\n";
-      return false;
-   }
-   return true;
-}
-
-bool CheckNontrivial( const char * label, const Real value )
-{
-   if ( !std::isfinite( value ) || value <= 1e-8 )
-   {
-      std::cout << "FAILED: " << label
-                << " produced a near-zero oracle, which would not detect "
-                << "duplicate inactive-lane contributions reliably.\n";
-      return false;
-   }
-   return true;
-}
-
-template < typename Config >
-bool CheckZeroWorkItems( const char * label )
-{
-   DeviceBuffer< Real > marker( 1, sentinel );
-   auto marker_data = marker.data;
-
-   Config::BlockLoop(
-      0,
-      [=] GENDIL_HOST_DEVICE ( Config ) mutable
-      {
-         marker_data[ 0 ] = 1.0;
-      } );
-   GENDIL_DEVICE_SYNC;
-
-   marker.CopyToHost();
-
-   const bool success = marker.data.host_pointer[ 0 ] == sentinel;
-   if ( !success )
-   {
-      std::cout << "FAILED: zero-work-item launch wrote in "
-                << label << ".\n";
-   }
-   return success;
-}
-
 template <
    typename Layout,
    Integer MaxSharedDimensions,
@@ -221,9 +81,15 @@ bool RunL2ErrorCaseForCellCount(
    if ( num_cells == 0 )
    {
       bool success = true;
-      success = CheckZeroWorkItems< LegacyConfig >( label ) && success;
-      success = CheckZeroWorkItems< DeviceBatch1 >( label ) && success;
-      success = CheckZeroWorkItems< DeviceBatchN >( label ) && success;
+      success =
+         CheckZeroWorkItems< LegacyConfig >( label, real_sentinel ) &&
+         success;
+      success =
+         CheckZeroWorkItems< DeviceBatch1 >( label, real_sentinel ) &&
+         success;
+      success =
+         CheckZeroWorkItems< DeviceBatchN >( label, real_sentinel ) &&
+         success;
       return success;
    }
 
@@ -253,21 +119,22 @@ bool RunL2ErrorCaseForCellCount(
    bool success = true;
    std::cout << label << ", num_cells = " << num_cells << '\n';
    success =
-      CheckNontrivial( "LegacyConfig L2Error", error_legacy ) && success;
+      CheckNontrivialPositive( "LegacyConfig L2Error", error_legacy ) &&
+      success;
    success =
-      CheckClose(
+      CheckScalarClose(
          "DeviceBatchN vs LegacyConfig",
          error_batchn,
          error_legacy,
          tolerance ) && success;
    success =
-      CheckClose(
+      CheckScalarClose(
          "DeviceBatchN vs DeviceBatch1",
          error_batchn,
          error_batch1,
          tolerance ) && success;
    success =
-      CheckClose(
+      CheckScalarClose(
          "DeviceBatch1 vs LegacyConfig",
          error_batch1,
          error_legacy,
@@ -296,8 +163,12 @@ bool RunRegisterOnlyL2ErrorCaseForCellCount(
    if ( num_cells == 0 )
    {
       bool success = true;
-      success = CheckZeroWorkItems< DeviceBatch1 >( label ) && success;
-      success = CheckZeroWorkItems< DeviceBatchN >( label ) && success;
+      success =
+         CheckZeroWorkItems< DeviceBatch1 >( label, real_sentinel ) &&
+         success;
+      success =
+         CheckZeroWorkItems< DeviceBatchN >( label, real_sentinel ) &&
+         success;
       return success;
    }
 
@@ -324,47 +195,12 @@ bool RunRegisterOnlyL2ErrorCaseForCellCount(
    constexpr Real tolerance = 1e-10;
    std::cout << label << ", num_cells = " << num_cells << '\n';
    return
-      CheckNontrivial( "DeviceBatch1 L2Error", error_batch1 ) &&
-      CheckClose(
+      CheckNontrivialPositive( "DeviceBatch1 L2Error", error_batch1 ) &&
+      CheckScalarClose(
          "Register-only DeviceBatchN vs DeviceBatch1",
          error_batchn,
          error_batch1,
          tolerance );
-}
-
-template < Integer BatchSize, typename Lambda >
-bool RunNormalizedCellCases( Lambda && run_case )
-{
-   const std::array< GlobalIndex, 6 > candidates{
-      0,
-      1,
-      BatchSize - 1,
-      BatchSize,
-      BatchSize + 1,
-      10
-   };
-   std::array< GlobalIndex, 6 > cases{};
-   Integer num_cases = 0;
-
-   for ( const GlobalIndex candidate : candidates )
-   {
-      bool found = false;
-      for ( Integer i = 0; i < num_cases; ++i )
-      {
-         found = found || cases[ i ] == candidate;
-      }
-      if ( !found )
-      {
-         cases[ num_cases++ ] = candidate;
-      }
-   }
-
-   bool success = true;
-   for ( Integer i = 0; i < num_cases; ++i )
-   {
-      success = run_case( cases[ i ] ) && success;
-   }
-   return success;
 }
 
 template <
